backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
pydantic==2.9.2
groq==0.11.0
web3==7.3.0
python-dotenv==1.0.1
aiohttp==3.10.5
aiofiles==24.1.0
httpx==0.27.2
